AvilaOps/products/windows-dev-optimizer/requirements_framework.txt
# Windows Dev Optimizer Framework Requirements
# Privacy-First FastAPI + HTMX + ON Platform Integration

# Core Web Framework
fastapi>=0.104.1,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
jinja2>=3.1.2,<4.0.0
python-multipart>=0.0.6,<1.0.0

# Security & Privacy
cryptography>=41.0.0,<42.0.0
python-jose[cryptography]>=3.3.0,<4.0.0

# Database & Storage
# sqlite3 is built into Python
aiofiles>=23.2.1,<24.0.0

# HTTP & Networking
httpx>=0.25.0,<1.0.0
websockets>=12.0,<13.0
pywebview>=4.4,<5.0

# Data Processing
pydantic>=2.4.0,<3.0.0
pydantic-settings>=2.0.3,<3.0.0

# Development Tools
rich>=13.6.0,<14.0.0
click>=8.1.7,<9.0.0

# Async Support
asyncio-mqtt>=0.14.0,<1.0.0  # If MQTT needed for ON platform

# Existing Windows Optimization Dependencies
psutil>=5.9.0,<6.0.0
# winreg, subprocess, json, pathlib, datetime, typing are built into Python

# Optional: Performance Monitoring
psutil>=5.9.0,<6.0.0
py-cpuinfo>=9.0.0,<10.0.0

# Optional: Advanced Analytics
pandas>=2.1.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Development Only
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
black>=23.9.0,<24.0.0
mypy>=1.5.0,<2.0.0